Reorganize Client/ by bindings
