+++ CMakeLists.txt
@@ -203,10 +203,6 @@ add_subdirectory(ChronoKeeper)
 add_subdirectory(ChronoGrapher)
 add_subdirectory(ChronoPlayer)
 add_subdirectory(ChronoStore)
-if(CHRONOLOG_ENABLE_PYTHON_BINDINGS)
-message("Python binding enabled")
-add_subdirectory(python_client)
-endif()
 if(CHRONOLOG_BUILD_TESTING)
     enable_testing()
     add_subdirectory(test)
