+++ Client/cpp/include/ClientConfiguration.h
