+++ Client/cpp/src/ClientQueryService.cpp
