+++ Client/cpp/src/ClientQueryService.h
