+++ Client/cpp/src/KeeperRecordingClient.h
