+++ Client/cpp/src/PlaybackQueryRpcClient.cpp
