+++ Client/cpp/src/PlaybackQueryRpcClient.h
