+++ Client/cpp/src/StorytellerClient.h
@@ -9,7 +9,7 @@
 #include "chrono_monitor.h"
 #include "KeeperIdCard.h"
 #include "chronolog_types.h"
-#include "chronolog_client.h"
+#include "../include/chronolog_client.h"
 
 #include "ClientQueryService.h"
