+++ Client/python/py_chronolog_client.cpp
@@ -2,7 +2,7 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
 
-#include <chronolog_client.h>
+#include "../cpp/include/chronolog_client.h"
